requirements_llava.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.32.0=pypi_0
aiofiles=23.2.1=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.10.0=pypi_0
attrs=25.3.0=pypi_0
autoattack=0.1=pypi_0
bitsandbytes=0.45.5=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
click=8.3.0=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
diffusers=0.27.2=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
exceptiongroup=1.3.0=pypi_0
expat=2.7.1=h6a678d5_0
fastapi=0.117.1=pypi_0
ffmpy=0.6.1=pypi_0
filelock=3.19.1=pypi_0
fonttools=4.60.0=pypi_0
fsspec=2025.9.0=pypi_0
gradio=4.16.0=pypi_0
gradio-client=0.8.1=pypi_0
h11=0.14.0=pypi_0
hf-xet=1.1.10=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.0=pypi_0
huggingface-hub=0.25.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.5.2=pypi_0
inflection=0.5.1=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.2=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
kiwisolver=1.4.9=pypi_0
latex2mathml=3.78.1=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
llava=1.2.2.post1=pypi_0
markdown-it-py=4.0.0=pypi_0
markdown2=2.5.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.6=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
narwhals=2.5.0=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.4.2=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.0.17=h5eee18b_0
orjson=3.11.3=pypi_0
packaging=25.0=pypi_0
pandas=2.3.2=pypi_0
peft=0.10.0=pypi_0
pillow=10.4.0=pypi_0
pip=25.2=pyhc872135_0
protobuf=6.32.1=pypi_0
psutil=7.1.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pycocotools=2.0.10=pypi_0
pydantic=2.11.9=pypi_0
pydantic-core=2.33.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.2.5=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.3=hc2a1206_0
referencing=0.36.2=pypi_0
regex=2025.9.18=pypi_0
requests=2.32.5=pypi_0
rich=14.1.0=pypi_0
rpds-py=0.27.1=pypi_0
ruff=0.13.1=pypi_0
safetensors=0.6.2=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.15.3=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=78.1.1=py310h06a4308_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.50.2=hb25bd0a_1
starlette=0.48.0=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.14.0=pypi_0
threadpoolctl=3.6.0=pypi_0
timm=0.6.13=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.15.1=pypi_0
tomlkit=0.12.0=pypi_0
torch=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.37.2=pypi_0
triton=2.1.0=pypi_0
typer=0.19.1=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.5.0=pypi_0
uvicorn=0.36.0=pypi_0
wavedrom=2.0.3.post3=pypi_0
websockets=11.0.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xtransferbench=0.1.2=pypi_0
xz=5.6.4=h5eee18b_1
zipp=3.23.0=pypi_0
zlib=1.3.1=hb25bd0a_0
